Inlined the RemoteBroker implementation, still not fully read but failures have reduced to 20 failing tests (from its previous 40 tests)
